site: split configs from users controller
To transform settings in menu with submenus
