+++ core/Language/Language.csproj
@@ -21,6 +21,7 @@
     <None Remove="Language\Email\WipeNotice.json" />
     <None Remove="Language\Site\accounts.json" />
     <None Remove="Language\Site\categories.json" />
+    <None Remove="Language\Site\configs.json" />
     <None Remove="Language\Site\email.json" />
     <None Remove="Language\Site\error.json" />
     <None Remove="Language\Site\general.json" />
@@ -101,6 +102,9 @@
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
     <Content Include="Language\Site\tokens.json">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="Language\Site\configs.json">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
     <Content Include="Language\Site\users.json">
