+++ core/Language/Language/Site/configs.json
@@ -1,0 +1,104 @@
+{
+	"en-US": {
+		"Save": "Save",
+		"MainSettings": "Main",
+		"PasswordSettings": "Password",
+		"EmailSettings": "E-mail",
+		"ThemeSettings": "Colors",
+		"TFASettings": "Safer login",
+		"TimeZone": "Time zone",
+		"UseCategories": "Use categories",
+		"SendMoveEmail": "Send e-mail when record move",
+		"MoveCheck": "Move can be checked / unchecked",
+		"ConfigChanged": "Settings changed",
+		"CurrentPassword": "Current password",
+		"NewPassword": "New password",
+		"EmailToChange": "New e-mail",
+		"ThemeInstruction": "Chose colors style and verify the examples below. Save to use the chosen colors at the website for now on.",
+		"ThemeLink": "These themes are based on {{link}}.",
+
+		"None": "OUCH!",
+		"Dark": "Dark",
+		"Light": "Light",
+		"Magic": "Magic",
+		"Sober": "Sober",
+		"Nature": "Nature",
+		"DarkMagic": "Magic (Dark)",
+		"DarkSober": "Sober (Dark)",
+		"DarkNature": "Nature (Dark)",
+		"LightMagic": "Magic (Light)",
+		"LightSober": "Sober (Light)",
+		"LightNature": "Nature (Light)",
+
+		"Alert": "Alert",
+		"Button": "Button",
+		"Active": "Active",
+		"Header": "Column",
+		"Label": "Field",
+		"RowNegative": "Negative Value",
+		"RowPositive": "Positive Value",
+		"RowTransfer": "Transfer Value",
+
+		"EmailUpdated": "E-mail successfully updated. New user activation is required.",
+		"PasswordChanged": "Password successfully updated. System logged out for safety.",
+
+		"TFAWhy": "Why to use this?",
+		"TFATitle": "This Configuration makes your account safer",
+		"TFAExplanation": "Each time one try to login, after typing e-mail and password, a code will be asked. This code is generated by cellphone apps. So, if someone break your password, it will still need to have your device to login.{0}To activate it, install an two-factor authentication app (we recommend using {1} or {2}), scan the QRCode below, and type the code generated here.",
+		"TFARemove": "Remove safer login",
+		"TFAPasswordEnable": "Use safer login\ncode as password",
+		"TFAPasswordDisable": "Stop using safer login\ncode as password",
+		"TFAAuthenticated": "Login mais seguro configurado!"
+	},
+	"pt-BR": {
+		"Save": "Salvar",
+		"MainSettings": "Geral",
+		"PasswordSettings": "Senha",
+		"EmailSettings": "E-mail",
+		"ThemeSettings": "Cores",
+		"TFASettings": "Login mais seguro",
+		"TimeZone": "Fuso horário",
+		"UseCategories": "Usar categorias",
+		"SendMoveEmail": "Enviar e-mail ao gravar movimentação",
+		"MoveCheck": "Movimentação pode ser verificada",
+		"ConfigChanged": "Configurações alteradas",
+		"CurrentPassword": "Senha atual",
+		"NewPassword": "Nova senha",
+		"EmailToChange": "Novo e-mail",
+		"ThemeInstruction": "Escolha um estilo e verifique os exemplos abaixo. Salve para usar no site de agora em diante o estilo que escolher.",
+		"ThemeLink": "Esses temas são baseados em {{link}}.",
+
+		"None": "EITA!",
+		"Dark": "Noite",
+		"Light": "Dia",
+		"Magic": "Magia",
+		"Sober": "Grave",
+		"Nature": "Natureza",
+		"DarkMagic": "Magia (Noite)",
+		"DarkSober": "Grave (Noite)",
+		"DarkNature": "Natureza (Noite)",
+		"LightMagic": "Magia (Dia)",
+		"LightSober": "Grave (Dia)",
+		"LightNature": "Natureza (Dia)",
+
+		"Alert": "Alerta",
+		"Button": "Botão",
+		"Active": "Ativo",
+		"Header": "Coluna",
+		"Label": "Campo",
+		"RowNegative": "Valor Negativo",
+		"RowPositive": "Valor Positivo",
+		"RowTransfer": "Valor Transferência",
+
+		"EmailUpdated": "E-mail atualizado com sucesso. Será necessária nova ativação no sistema.",
+		"PasswordChanged": "Senha atualizada com sucesso. Sistema deslogado por segurança.",
+
+		"TFAWhy": "Por que usar?",
+		"TFATitle": "Esta configuração mantém sua conta mais segura",
+		"TFAExplanation": "A cada vez que for se logar, depois de digitar e-mail e senha, um código será pedido. Esse código é gerado por aplicativos de celular. Então, se alguém descobrir sua senha, ainda vai precisar ter acesso ao seu dispositivo para entrar.{0}Para ativar, instale um aplicativo de autenticação em duas etapas (recomendamos {1} ou {2}), aponte para o QRCode, e digite aqui o código que aparecer no aplicativo.",
+		"TFARemove": "Remover login mais seguro",
+		"TFAPasswordEnable": "Usar o código de login\nmais seguro como senha",
+		"TFAPasswordDisable": "Deixar de usar o código de\nlogin mais seguro como senha",
+		"TFAAuthenticated": "Safer login set!"
+	}
+}
