+++ core/Language/Language/Site/tokens.json
@@ -5,7 +5,6 @@
 		"TokenInvalid": "Token Invalid",
 		"TokenInvalidMessage": "The Token is invalid.",
 		"Password": "Password",
-		"RetypePassword": "Retype password",
 		"Reset": "Reset",
 		"PasswordResetSuccess": "Password successfully reset.",
 		"Token": "Token",
@@ -21,7 +20,6 @@
 		"TokenInvalid": "Token Inválido",
 		"TokenInvalidMessage": "Este Token é inválido.",
 		"Password": "Senha",
-		"RetypePassword": "Redigite a senha",
 		"Reset": "Redefinir",
 		"PasswordResetSuccess": "Senha redefinida com sucesso.",
 		"Token": "Token",
