+++ core/Language/Language/Site/users.json
@@ -2,7 +2,6 @@
 	"en-US": {
 		"Email": "E-mail",
 		"Password": "Password",
-		"RetypePassword": "Retype password",
 		"SignUpSuccess": "Sign up succeeded.",
 		"LogonDisabled": "Your user is not active.",
 		"RememberMe": "Remember",
@@ -10,65 +9,17 @@
 		"EmailSentIfUserExists": "If the user exists, you will receive an e-mail with instructions.",
 		"FollowEmailInstructions": "Follow the instructions in the e-mail sent.",
 		"PayAttentionSender": "Pay attention to the sender address: it MUST be ",
-		"UseCategories": "Use categories",
-		"SendMoveEmail": "Send e-mail when record move",
-		"MoveCheck": "Move can be checked / unchecked",
-		"TimeZone": "Time zone",
-		"Save": "Save",
-		"ConfigChanged": "Settings changed",
-		"CurrentPassword": "Current password",
-		"NewPassword": "New password",
-		"EmailToChange": "New e-mail",
-		"EmailUpdated": "E-mail successfully updated. New user activation is required.",
-		"PasswordChanged": "Password successfully updated. System logged out for safety.",
-		"MainSettings": "Main",
-		"PasswordSettings": "Password",
-		"EmailSettings": "E-mail",
-		"ThemeSettings": "Colors",
-		"TFASettings": "Safer login",
-		"EnableTFA": "Enable two-factor",
-		"TFAAuthenticated": "Safer login set!",
-		"TFAWhy": "Why to use this?",
-		"TFARemove": "Remove safer login",
-		"TFAContactUs": "I don't have the code",
-		"TFAPasswordEnable": "Use safer login\ncode as password",
-		"TFAPasswordDisable": "Stop using safer login\ncode as password",
+		"TFA": "Safer login",
 
-		"Alert": "Alert",
-		"Button": "Button",
-		"Active": "Active",
-		"Header": "Column",
-		"Label": "Field",
-		"RowNegative": "Negative Value",
-		"RowPositive": "Positive Value",
-		"RowTransfer": "Transfer Value",
-
-		"Code": "Code",
 		"ContractAccept": "I understood and agree with the {0}",
 		"ContractBeginDate": "Effective Date",
 		"CookiesWarning": "Cookies are used to control logon and security of forms of this website",
-		"EnableWizard": "Enable Wizard",
 		"TFAVerification": "Use the code of the app registered before",
-		"ThemeInstruction": "Chose colors style and verify the examples below. Save to use the chosen colors at the website for now on.",
-		"ThemeLink": "These themes are based on {{link}}.",
-
-		"None": "OUCH!",
-		"Dark": "Dark",
-		"Light": "Light",
-		"Magic": "Magic",
-		"Sober": "Sober",
-		"Nature": "Nature",
-		"DarkMagic": "Magic (Dark)",
-		"DarkSober": "Sober (Dark)",
-		"DarkNature": "Nature (Dark)",
-		"LightMagic": "Magic (Light)",
-		"LightSober": "Sober (Light)",
-		"LightNature": "Nature (Light)",
+		"TFAContactUs": "I don't have the code",
 	},
 	"pt-BR": {
 		"Email": "E-mail",
 		"Password": "Senha",
-		"RetypePassword": "Redigite a senha",
 		"SignUpSuccess": "Cadastro efetuado com sucesso.",
 		"LogonDisabled": "Seu acesso ao sistema não foi ativado ainda.",
 		"RememberMe": "Lembrar",
@@ -76,59 +27,12 @@
 		"EmailSentIfUserExists": "Se existir este e-mail no sistema, você receberá as instruções para prosseguir.",
 		"FollowEmailInstructions": "Siga as instruções no e-mail enviado.",
 		"PayAttentionSender": "Atenção para remetente de e-mail: PRECISA ser ",
-		"UseCategories": "Usar categorias",
-		"SendMoveEmail": "Enviar e-mail ao gravar movimentação",
-		"MoveCheck": "Movimentação pode ser verificada",
-		"TimeZone": "Fuso horário",
-		"Save": "Salvar",
-		"ConfigChanged": "Configurações alteradas",
-		"CurrentPassword": "Senha atual",
-		"NewPassword": "Nova senha",
-		"EmailToChange": "Novo e-mail",
-		"EmailUpdated": "E-mail atualizado com sucesso. Será necessária nova ativação no sistema.",
-		"PasswordChanged": "Senha atualizada com sucesso. Sistema deslogado por segurança.",
-		"MainSettings": "Geral",
-		"PasswordSettings": "Senha",
-		"EmailSettings": "E-mail",
-		"ThemeSettings": "Cores",
-		"TFASettings": "Login mais seguro",
-		"EnableTFA": "Habilitar segundo fator",
-		"TFAAuthenticated": "Login mais seguro configurado!",
-		"TFAWhy": "Por que usar?",
-		"TFARemove": "Remover login mais seguro",
-		"TFAContactUs": "não sei o código",
-		"TFAPasswordEnable": "Usar o código de login\nmais seguro como senha",
-		"TFAPasswordDisable": "Deixar de usar o código de\nlogin mais seguro como senha",
+		"TFA": "Login mais seguro",
 
-		"Alert": "Alerta",
-		"Button": "Botão",
-		"Active": "Ativo",
-		"Header": "Coluna",
-		"Label": "Campo",
-		"RowNegative": "Valor Negativo",
-		"RowPositive": "Valor Positivo",
-		"RowTransfer": "Valor Transferência",
-
-		"Code": "Código",
 		"ContractAccept": "Entendi e concordo com os {0}",
 		"ContractBeginDate": "Início de Vigência",
 		"CookiesWarning": "Cookies são usados para controlar seu acesso a informações e a segurança dos formulários deste site",
-		"EnableWizard": "Habilitar Tutorial",
 		"TFAVerification": "Use o código gerado pelo aplicativo cadastrado anteriormente",
-		"ThemeInstruction": "Escolha um estilo e verifique os exemplos abaixo. Salve para usar no site de agora em diante o estilo que escolher.",
-		"ThemeLink": "Esses temas são baseados em {{link}}.",
-
-		"None": "EITA!",
-		"Dark": "Noite",
-		"Light": "Dia",
-		"Magic": "Magia",
-		"Sober": "Grave",
-		"Nature": "Natureza",
-		"DarkMagic": "Magia (Noite)",
-		"DarkSober": "Grave (Noite)",
-		"DarkNature": "Natureza (Noite)",
-		"LightMagic": "Magia (Dia)",
-		"LightSober": "Grave (Dia)",
-		"LightNature": "Natureza (Dia)",
+		"TFAContactUs": "não sei o código",
 	}
 }
