+++ core/Language/Language/Site/wizard.json
@@ -9,8 +9,8 @@
 		"Wizard_EndTextAndroid": "Don't fly Money has an android app available at Play Store:",
 		"Wizard_EndTextContact": "If you have something to tell us, you also can send an e-mail to {0}. We will answer as soon as possible.",
 		"Wizard_EndTextMenu": "Just some last things to see on top menu!",
-		"Wizard_EndTextMenuSettings": "Re-enable the Wizard by going to 'Settings'",
-		"Wizard_EndTextMenuContact": "Contact us and see more about the project by going to 'Contact'",
+		"Wizard_EndTextMenuSettings": "Re-enable the Wizard at 'Settings' > 'Main'",
+		"Wizard_EndTextMenuContact": "Contact us and see more about the project at 'Contact'",
 		"Wizard_EndTextThanks": "Thanks for reading our wizard messages.\nWe hope now you know Don't fly Money better.",
 		"Wizard_Logins": "Here is a security area: you can see and disable system logins, accesses to your registries. If something seems suspect or some device logged in was accessed by other people without authorization, we suggest you take down the Logins, and the next request will need to know the login credentials to enter again.",
 		"Wizard_MovesCreate": "A Move is when the money gets In or Out of the Account - and it can be transferred from/to another one. To add it, give it a Description, a Date and a Value. It's also possible, instead of use a single Value, to create it with Details.\n\nREMEMBER: THE SYSTEM HAS NO CONNECTION TO ANY FINANCIAL INSTITUTION, SO ANY MOVES MADE HERE ARE MADE JUST HERE, BEING MERELY REGISTER IN THIS SYSTEM.",
@@ -33,8 +33,8 @@
 		"Wizard_EndTextAndroid": "Don't fly Money tem um aplicativo android na Play Store:",
 		"Wizard_EndTextContact": "Se tiver algo a nos dizer, também pode entrar em contato pelo e-mail {0}. Responderemos assim que possível.",
 		"Wizard_EndTextMenu": "Apenas algumas últimas coisas que pode encontrar no menu!",
-		"Wizard_EndTextMenuSettings": "Reabilite o Tutorial em 'Configurações'",
-		"Wizard_EndTextMenuContact": "Entre em contato e veja mais sobre o projeto em 'Contato'",
+		"Wizard_EndTextMenuSettings": "Reabilite o Tutorial em 'Configurações' > 'Geral'",
+		"Wizard_EndTextMenuContact": "Fale conosco e veja mais sobre o projeto em 'Contato'",
 		"Wizard_EndTextThanks": "Agradecemos por ler as mensagens do Tutorial.\nEsperamos que agora conheça melhor o sistema Don't fly Money.",
 		"Wizard_Logins": "Aqui você pode ver e desabilitar Logins feitos no sistema. Se desconfiar de algum acesso, ou algum dispositivo logado tenha sido acessado por pessoas não autorizadas, é sugerido que os Logins sejam excluídos. Isso impedirá novo acesso sem uso de senha.",
 		"Wizard_MovesCreate": "Uma Movimentação é uma Saída, Entrada ou Transferência de dinheiro. Para inserir, dê uma Descrição, uma Data e o Valor. Também é possível, ao invés de inserir um Valor único, inserir com Detalhes.\n\nLEMBRANDO QUE O SISTEMA NÃO POSSUI LIGAÇÃO COM NENHUMA INSTITUIÇÃO EXTERNA, PORTANTO MOVIMENTAÇÕES FEITAS AQUI SÃO APENAS REGISTROS FEITOS POR VOCÊ INTERNAMENTE NESTE SISTEMA.",
