+++ site/MVC/Models/Configs/ThemeOptions.cs
@@ -5,7 +5,7 @@ using DFM.BusinessLogic.Services;
 using DFM.Generic;
 using DFM.MVC.Helpers.Global;
 
-namespace DFM.MVC.Models.UserConfig
+namespace DFM.MVC.Models.Configs
 {
 	public class ThemeOptions
 	{
