+++ site/MVC/Models/UsersTFAModel.cs
@@ -20,10 +20,5 @@ namespace DFM.MVC.Models
 				);
 			}
 		}
-
-		public void UseAsPassword(Boolean use)
-		{
-			safe.UseTFAAsPassword(use);
-		}
 	}
 }
