+++ site/MVC/Views/Configs/ConfigEmail.cshtml
@@ -1,4 +1,4 @@
-﻿@model UsersConfigModel
+﻿@model ConfigsConfigModel
 
 <div class="form-group">
 	@Html.LabelFor(m => m.Info.Email, Context.Translate("EmailToChange"), new { @class = "control-label" })
