+++ site/MVC/Views/Configs/ConfigMain.cshtml
@@ -1,4 +1,4 @@
-﻿@model UsersConfigModel
+﻿@model ConfigsConfigModel
 
 <div class="form-group">
 	@Html.LabelFor(m => m.Main.Language, Context.Translate("Language"), new { @class = "control-label" })
