+++ site/MVC/Views/Configs/ConfigPassword.cshtml
@@ -1,4 +1,4 @@
-﻿@model UsersConfigModel
+﻿@model ConfigsConfigModel
 
 <div class="form-group">
 	@Html.LabelFor(m => m.Info.Password.CurrentPassword, Context.Translate("CurrentPassword"), new { @class = "control-label" })
