+++ site/MVC/Views/Configs/ConfigTFA.cshtml
@@ -1,4 +1,4 @@
-﻿@model UsersConfigModel
+﻿@model ConfigsConfigModel
 
 @Html.HiddenFor(m => m.TFA.IsActive)
