+++ site/MVC/Views/Configs/ConfigTheme.cshtml
@@ -1,4 +1,4 @@
-﻿@model UsersConfigModel
+﻿@model ConfigsConfigModel
 
 @Html.HiddenFor(m => m.ThemeOpt.Theme, new { @class = "theme-changer-hidden" })
