--- site/MVC/Views/Shared/Skeleton/MenuSettings.cshtml
@@ -1,7 +1,0 @@
-﻿@{
-	var route = new Default.Main();
-}
-
-<a href="@Url.RouteUrl(route.Name, new {action = "Config", controller = "Users", area = route.Area})" role="button">
-	<span class="glyphicon glyphicon-cog" title="@Context.Translate("Settings")"></span>
-</a>
