+++ site/MVC/Views/Shared/Skeleton/Wizard.cshtml
@@ -18,7 +18,7 @@
 			
 			@await Html.PartialAsync("Extensions/MicroForm",
 				MicroForm.WithResource(Context, Context.Translate("Wizard_End"), "btn btn-sm btn-info pull-right")
-					.AddRouteUrl<Default.Main>("Users", "EndWizard")
+					.AddRouteUrl<Default.Main>("Configs", "EndWizard")
 			)
 		</div>
