+++ site/MVC/Views/Tokens/UnsubscribeMoveMailSuccess.cshtml
@@ -10,6 +10,6 @@
 
 @Html.ActionLink(
 	Context.Translate("GoToSettings"),
-	"Config", "Users",
+	"Config", "Configs",
 	null, new { @class = "alert-link" }
 )
