+++ site/MVC/Views/Users/TFA.cshtml
@@ -2,7 +2,7 @@
 
 @{
 	Layout = "~/Views/Shared/Layouts/_Form.cshtml";
-	ViewBag.Title = Context.Translate("TFASettings");
+	ViewBag.Title = Context.Translate("TFA");
 }
 
 <div class="form-group">
